New smell system

+++ AntSim/Simulation/Map/Cell.cs
@@ -6,6 +6,6 @@ namespace AntSim.Simulation.Map
     class Cell
     {
         public GraphicalObject Entity { get; set; } = null;
-        public Dictionary<byte, uint> Smells { get; set; } = new Dictionary<byte, uint>();
+        public Dictionary<uint, uint> Smells { get; set; } = new Dictionary<uint, uint>();
     }
 }
